requirements.txt
Pillow~=9.3.0
tqdm~=4.64.1